DEV Move SparseCSCStreamProcessor to s/p/base.py
